+++ headers/vector.h
@@ -5,14 +5,14 @@ typedef struct vec_2d vec_2d;
 
 struct vec_2d
 {
-	int x, y;
+	float x, y;
 };
 
 typedef struct vec_3d vec_3d;
 
 struct vec_3d
 {
-	int x, y, z;
+	float x, y, z;
 };
 
 #endif
